+++ R/globals.R
@@ -2,7 +2,7 @@
 utils::globalVariables(".")
 utils::globalVariables("HLAalignments")
 utils::globalVariables("alleleListHistory")
-utils::globalVariables("HLAgazeteer")
+utils::globalVariables("HLAgazetteer")
 utils::globalVariables("fragmentFeatureNames")
 utils::globalVariables("IMGTHLAGeneTypes")
 utils::globalVariables("HLAatlas")